refactor: merge to simplify structure
