+++ src/generators/index.ts
@@ -1,7 +1,6 @@
 import { Generator } from "../generator";
-import jsdocs from "./jsdocs";
-import pmInstall from "./pm/install";
-import pmX from "./pm/x";
+import { jsdocs } from "./jsdocs";
+import { pmX, pmInstall } from "./pm";
 
 export default {
   jsdocs,
