+++ src/generators/jsdocs.ts
@@ -4,7 +4,7 @@ import type { Schema } from "untyped";
 import { titleCase } from "scule";
 import { defineGenerator } from "../generator";
 
-export default defineGenerator({
+export const jsdocs = defineGenerator({
   name: "jsdocs",
   async generate({ options, args }) {
     const entryPath = resolve(options.dir, args.src || "./src/index");
